+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog for EtwinBar
 This changelog is formatted as recommended by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+### Added
+- Can insert the full footer of Eternaltwin (with the Piouz logo, "Thanks to" block and "Devs" block)
+
 ## [1.1] - 2025-07-30
 ### Added
 - [a11y] Mention that the links open in a new tab